+++ src/core/utils/fetchers/POP3_SSL_Fetcher.py
@@ -56,13 +56,11 @@ class POP3_SSL_Fetcher(  # noqa: N801  # naming consistent with POP3_SSL class
         mail_host = self.account.mail_host
         mail_host_port = self.account.mail_host_port
         timeout = self.account.timeout
-        ssl_context = ssl.create_default_context(
-            purpose=(
-                ssl.Purpose.CLIENT_AUTH
-                if get_config("ALLOW_INSECURE_CONNECTIONS")
-                and self.account.allow_insecure_connection
-                else ssl.Purpose.SERVER_AUTH
-            )
+        ssl_context = (
+            ssl._create_unverified_context()  # noqa: SLF001 S323  # pylint: disable=protected-access  # method is private due to its risk to security
+            if get_config("ALLOW_INSECURE_CONNECTIONS")
+            and self.account.allow_insecure_connection
+            else ssl.create_default_context()
         )
         try:
             if mail_host_port: